Fixed: Corrected spell_start + spell_go packets, a non ingame-visible patch. Patch provide by ratatatata. Fixed: problem with freeze mangosd at transports loading. My typo in previous patch.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2717 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Transports.cpp
@@ -210,7 +210,7 @@ bool Transport::GenerateWaypoints(uint32 pathid, vector <uint32> &mapids)
         keyFrames[j].distSinceStop = tmpDist;
     }
 
-    for (size_t i = keyFrames.size() - 1; i >= 0; i--) {
+    for (int i = int(keyFrames.size()) - 1; i >= 0; i--) {
         int j = (i + (keyFrames.size() - lastStop)) % keyFrames.size();
         tmpDist += keyFrames[(j + 1) % keyFrames.size()].distFromPrev;
         keyFrames[j].distUntilStop = tmpDist;
